+++ deploy/gke/autopush.yaml
@@ -23,3 +23,7 @@ nodes: 3
 storage_project: datcom-store
 model_bucket: "gs://datcom-ipragma-models"
 model_name: "dc_t5_ir3_fix_20220811_default_bs1bm1_gpu_jit1"
+
+website:
+  placeSummary:
+    enabled: true
