+++ scripts/deploy_gke_helm.sh
@@ -148,7 +148,12 @@ function deploy_website() {
   --set cronTesting.nodejsApiRoot="$NODEJS_SERVICE_URL" \
   --set-file nl.embeddings=deploy/nl/embeddings.yaml \
   --set-file nl.models=deploy/nl/models.yaml \
-  --set-file website.placeSummary=server/config/summaries/place_summaries.json
+  --set-file website.placeSummary.data.country_=server/config/place_summaries_for_country_.json \
+  --set-file website.placeSummary.data.geoId_0_2=server/config/place_summaries_for_geoId_0-2.json \
+  --set-file website.placeSummary.data.geoId_3_5=server/config/place_summaries_for_geoID_3-5.json \
+  --set-file website.placeSummary.data.geoId_6_9=server/config/place_summaries_for_geoId_6-9.json \
+  --set-file website.placeSummary.data.wikidataId_=server/config/place_summaries_for_wikidataId_.json \
+  --set-file website.placeSummary.data.others=server/config/place_summaries_others.json
 }
 
 cd $ROOT
